+++ packages/gatsby/src/commands/build.js
@@ -184,10 +184,13 @@ module.exports = async function build(program: BuildArgs) {
   }
   activity.done()
 
+  activity = report.activityTimer(`onPostBuild`, { parentSpan: buildSpan })
+  activity.start()
   await apiRunnerNode(`onPostBuild`, {
     graphql: graphqlRunner,
     parentSpan: buildSpan,
   })
+  activity.end()
 
   // Make sure we saved the latest state so we have all jobs cached
   await db.saveState()