+++ packages/gatsby/src/query/query-compiler.js
@@ -18,8 +18,6 @@ const {
   visitWithTypeInfo,
   TypeInfo,
   isAbstractType,
-  isObjectType,
-  isInterfaceType,
   Kind,
   FragmentsOnCompositeTypesRule,
   KnownTypeNamesRule,
@@ -497,7 +495,7 @@ const addExtraFields = (document, schema) => {
       [Kind.SELECTION_SET]: node => {
         // Entering selection set:
         //   selection sets can be nested, so keeping their metadata stacked
-        contextStack.push({ hasTypename: false, hasId: false })
+        contextStack.push({ hasTypename: false })
       },
       [Kind.FIELD]: node => {
         // Entering a field of the current selection-set:
@@ -508,9 +506,6 @@ const addExtraFields = (document, schema) => {
           node?.alias?.value === `__typename`
         ) {
           context.hasTypename = true
-        }
-        if (node.name.value === `id` || node?.alias?.value === `id`) {
-          context.hasId = true
         }
       },
     },
@@ -528,16 +523,6 @@ const addExtraFields = (document, schema) => {
             name: { kind: Kind.NAME, value: `__typename` },
           })
         }
-        if (
-          !context.hasId &&
-          (isObjectType(parentType) || isInterfaceType(parentType)) &&
-          hasIdField(parentType)
-        ) {
-          extraFields.push({
-            kind: Kind.FIELD,
-            name: { kind: Kind.NAME, value: `id` },
-          })
-        }
         return extraFields.length > 0
           ? { ...node, selections: [...extraFields, ...node.selections] }
           : undefined
@@ -546,10 +531,4 @@ const addExtraFields = (document, schema) => {
   })
 
   return visit(document, transformer)
-}
-
-const hasIdField = type => {
-  const idField = type.getFields()[`id`]
-  const fieldType = idField ? String(idField.type) : ``
-  return fieldType === `ID` || fieldType === `ID!`
 }