+++ packages/gatsby-source-mongodb/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.32](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-mongodb@2.1.31...gatsby-source-mongodb@2.1.32) (2020-02-13)
+
+**Note:** Version bump only for package gatsby-source-mongodb
+
 ## [2.1.31](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-mongodb@2.1.30...gatsby-source-mongodb@2.1.31) (2020-02-02)
 
 ### Bug Fixes